requirements.txt
PySide6>=6.6.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
scipy>=1.11.0
scikit-image>=0.21.0
openpyxl>=3.1.0
